pyproject.toml
[project]
name = "ml-project"
version = "0.1.0"
description = "ML Project for CV, PMLDL, and RL courses"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    "torch>=2.1.2",
    "torchvision>=0.16.2",
    "sqlalchemy[asyncio]>=2.0.43",
    "pyjwt>=2.10.1",
    "pwdlib[argon2]>=0.2.1",
    "uvicorn>=0.35.0",
    "aiosqlite>=0.21.0",
    "alembic>=1.16.5",
    "dotenv>=0.9.9",
    "asyncclick>=8.3.0.7",
    "click>=8.2.1",
    "pillow>=11.3.0",
    "pandas>=2.3.2",
    "tqdm>=4.67.1",
    "torchmetrics>=1.8.2",
    "facenet-pytorch>=2.5.3",
    "dvc[ssh]>=3.63.0",
    "dvclive>=3.48.5",
    "prometheus-client>=0.23.1",
    "opentelemetry-distro>=0.59b0",
    "opentelemetry-exporter-otlp>=1.38.0",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
    "opentelemetry-instrumentation-logging>=0.59b0",
    "asyncpg>=0.30.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "httpx>=0.28.1",
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.6",
    "seaborn>=0.13.2",
    "torchmetrics>=1.8.2",
    "mlflow>=3.4.0",
    "pytest>=8.4.2",
    "pytest-cases>=3.9.1",
    "ruff>=0.14.0",
    "pandas>=2.3.2",
    "aiosqlite>=0.21.0",
    "colorlog>=6.9.0",
    "dotenv>=0.9.9",
    "tqdm>=4.67.1",
    "pre-commit>=4.3.0",
    "psutil>=7.0.0",
    "pynvml>=13.0.1",
    "asyncclick>=8.3.0.7",
    "opencv-python>=4.12.0.88",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
